Rename Fuzz as Fuzziness
Somehow, this trait has strayed from its original purpose. We will repurpose it now.
